memoize methods passed to rte
