+++ packages/fondue/src/components/RichTextEditor/components/RenderPlaceholder/index.ts
@@ -1,0 +1,3 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export * from './RenderPlaceholder';
